requirements.txt
pynput
customtkinter
pyinstaller
keyboard
pillow
